Refactoring games hub files
